irework again - attempt to understand
When I got it working initially in the scales2 branch it was with a bad approach. I would pass in the float* to the bit-mapped scale data, and then create a uint8_t array for each scale inside the threaded deconvolution tools class.

This would be forced into a boolean pointer array before being passed down to the peak finding with mask method.

In cleaning this up and moving to precomputed masks based on an aocommon::UVector the behavour changed. It became a sitation where teh clean components were 'off' from the actual masked scales. You coiuld see this in visualisation tools like carta. It was almost like a shear. I suspect that maybe the wrong ptr++ size was being used somewhere in the peak finding. As in the pointer was not incrementing through the memory array appropriately.

I am attempting again to get this right.